chared: add pic bitmap fill mode and remove early return from grid each
if we set edit fill to a pixel value we expect the grid data to be bit
selection. this saves 93.75% in binary and at around 50% and commonly 75%
in json for common draw edits and we can avoid the hack for painting 0.

we can narrow down the grid with a box already. i cannot think of
common uses that would require early return other than to check
whether or not a pixel is on a sub grid. but that should not be often.